I rethink about playlist-music service, and add new class and service to keep SOLID, I mean 1 class one responsibility
