+++ src/main/java/music/musicapp/repository/MusicRepository.java
@@ -10,4 +10,6 @@ import java.util.List;
 
 public interface MusicRepository extends JpaRepository<Music, Long> {
     List<Music> findByTitleContaining(String query);
+
+    Music addMusicToPlaylist(Music music);
 }
