+++ src/main/java/music/musicapp/repository/PlaylistRepository.java
@@ -4,9 +4,8 @@ import music.musicapp.model.Playlist;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
 
 public interface PlaylistRepository extends JpaRepository<Playlist, Long> {
+
 }
